backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.44
python-dotenv==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.2.0
python-dateutil==2.8.2
psycopg2-binary==2.9.9
requests==2.31.0
